Minimum 24hr value now works for negatives
Since theres been so many negative prices lately it was getting a bit annoying that the minimum price wasn't working for that. 
It was a stupid error in the code, I was clipping them intentionally. Oh well. 
Now should show the proper minimum price in the next 24h, rather than the lowest positive value!